feat: added NotifyImageView and NotifyImageHolder
The NotifyImageView permits an Activity to be notified via callback
after an ImageView's image content changes. This is useful when loading
images from a background thread, for example via the picasso library.
This class allows some action to be taken once the image loads.

The NotifyImageHolder interface must be implemented to receive notification.